Add workflow to trigger a build of the package repo
This workflow is an adapted copy of this workflow :
https://github.com/qualcomm-linux/pkg-template/blob/main/.github/workflows/to_paste_in_upstream/pkg-build-pr-check.yml

With this workflow in position, a PR in the upstream project repo (this
one) triggers a build or the sister package repo. The special repo
variable PKG_REPO_GITHUB_NAME will need to be set.

See visit https://github.com/qualcomm-linux/pkg-template for more info
about this whole CI workflow

Signed-off-by: Simon Beaudoin <sbeaudoi@qti.qualcomm.com>